requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
abseil-cpp=20211102.0=hd638f34_2
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.4=py39hdc70f33_1
aiosignal=1.3.1=pyhd8ed1ab_0
async-timeout=4.0.2=py39hecd8cb5_0
attrs=23.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.6.2=pyhd8ed1ab_0
brotli=1.0.9=hb7f2c08_8
brotli-bin=1.0.9=hb7f2c08_8
brotlipy=0.7.0=py39h9ed2024_1003
bzip2=1.0.8=h1de35cc_0
c-ares=1.19.1=h0dc2134_0
ca-certificates=2023.5.7=h8857fd0_0
cachetools=5.3.0=pyhd8ed1ab_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py39h6c40b1e_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=unix_pyhd8ed1ab_2
cloudpickle=2.2.1=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.7=py39h92daf61_0
cryptography=39.0.1=py39hf6deb26_1
cycler=0.11.0=pyhd8ed1ab_0
ffmpeg=4.3=h0a44026_0
filelock=3.9.0=py39hecd8cb5_0
fonttools=4.39.4=py39hdc70f33_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.3.3=py39ha30fb19_0
future=0.18.3=pyhd8ed1ab_0
gettext=0.21.0=he85b6c0_1
giflib=5.2.1=h6c40b1e_3
gmp=6.2.1=he9d5cce_3
gmpy2=2.1.2=py39hd5de756_0
gnutls=3.6.15=hed9c0bf_0
google-auth=2.19.1=pyh1a96a4e_0
google-auth-oauthlib=1.0.0=pyhd8ed1ab_0
gpytorch=1.11=pypi_0
grpc-cpp=1.48.2=h3afe56f_0
grpcio=1.48.2=py39h3afe56f_0
hyperopt=0.2.7=pyhd8ed1ab_0
icu=58.2=h0a44026_3
idna=3.4=py39hecd8cb5_0
importlib-metadata=6.6.0=pyha770c72_0
importlib-resources=5.12.0=pyhd8ed1ab_0
importlib_resources=5.12.0=pyhd8ed1ab_0
intel-openmp=2023.1.0=ha357a0b_43547
jaxtyping=0.2.20=pypi_0
jinja2=3.1.2=py39hecd8cb5_0
joblib=1.2.0=pypi_0
jpeg=9e=h6c40b1e_1
kiwisolver=1.4.4=py39h92daf61_1
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libabseil=20211102.0=cxx17_hd8326b1_2
libbrotlicommon=1.0.9=hb7f2c08_8
libbrotlidec=1.0.9=hb7f2c08_8
libbrotlienc=1.0.9=hb7f2c08_8
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_0
libffi=3.4.4=hecd8cb5_0
libgfortran=5.0.0=11_3_0_h97931a8_31
libgfortran5=12.2.0=he409387_31
libiconv=1.16=hca72f7f_2
libidn2=2.3.4=h6c40b1e_0
libpng=1.6.39=h6c40b1e_0
libprotobuf=3.20.3=hfff2838_0
libtasn1=4.19.0=h6c40b1e_0
libtiff=4.5.0=hcec6c5f_2
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.4=hf6ce154_1
libwebp-base=1.2.4=h6c40b1e_1
libxml2=2.10.3=h930c0e2_0
linear-operator=0.5.0=pypi_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
markdown=3.4.3=pyhd8ed1ab_0
markupsafe=2.1.1=py39hca72f7f_0
matplotlib=3.7.1=py39h6e9494a_0
matplotlib-base=3.7.1=py39hb2f573b_0
mkl=2023.1.0=h59209a4_43558
mkl-service=2.4.0=py39h6c40b1e_1
mkl_fft=1.3.6=py39h07fba90_1
mkl_random=1.2.2=py39h07fba90_1
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.2=h9066e36_1
mpmath=1.2.1=py39hecd8cb5_0
multidict=6.0.4=py39ha30fb19_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=hcec6c5f_0
nettle=3.7.3=h230ac6f_1
networkx=2.8.4=py39hecd8cb5_1
numpy=1.24.3=py39h47b59a4_1
numpy-base=1.24.3=py39hcfaf2c3_1
oauthlib=3.2.2=pyhd8ed1ab_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1u=h8a1eda9_0
packaging=23.1=pyhd8ed1ab_0
pandas=2.0.2=pypi_0
pillow=9.4.0=py39hcec6c5f_0
pip=23.0.1=py39hecd8cb5_0
protobuf=3.20.3=py39h7a8716b_1
py4j=0.10.9.7=pyhd8ed1ab_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd3eb1b0_0
pyjwt=2.7.0=pyhd8ed1ab_0
pyopenssl=23.0.0=py39hecd8cb5_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.16=h218abb5_2
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=2.0.1=py3.9_0
pytz=2023.3=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
re2=2022.04.01=h96cf925_0
readline=8.2=hca72f7f_0
requests=2.29.0=py39hecd8cb5_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.2.2=pypi_0
scipy=1.9.3=py39hf241641_2
setuptools=67.8.0=py39hecd8cb5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h6c40b1e_0
sympy=1.11.1=py39hecd8cb5_0
tbb=2021.8.0=ha357a0b_0
tensorboard=2.13.0=pyhd8ed1ab_0
tensorboard-data-server=0.7.0=py39h7242b5c_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h5d9f67b_0
torch-tb-profiler=0.4.1=pypi_0
torchaudio=2.0.2=py39_cpu
torchvision=0.15.2=py39_cpu
tornado=6.3.2=py39hdc70f33_0
tqdm=4.65.0=pyhd8ed1ab_1
typeguard=2.13.3=pypi_0
typing_extensions=4.5.0=py39hecd8cb5_0
tzdata=2023.3=pypi_0
unicodedata2=15.0.0=py39ha30fb19_0
urllib3=1.26.16=py39hecd8cb5_0
werkzeug=2.3.6=pyhd8ed1ab_0
wheel=0.38.4=py39hecd8cb5_0
xz=5.4.2=h6c40b1e_0
yarl=1.9.2=py39hdc70f33_0
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.5=hc035e20_0
